requirements.txt
# Core Flask and Web Framework
flask==3.0.0
flask-cors==4.0.0

# Data Processing and Analysis
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Optional: For Streamlit version (existing)
streamlit==1.29.0

# Utilities
python-dateutil==2.8.2

# AI Integration
google-generativeai==0.3.2
python-dotenv==1.0.0